effectif_*: mutualisation de periodCol, parseEffectifColMapping() et parseEffectifPeriod()
